requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
groq
aiofiles
python-dotenv